requirements.txt
flask==2.0.1
flask-cors==3.0.10
pillow==9.0.0
python-dotenv==0.19.1
requests==2.26.0
supabase==0.0.2
